Fix AXON Primary network interface issue
Network interface name is not same in windows machine.
It is different across different workloads. We can grep
by some string like 'Ethernet', but that will not work
when instance has multiple interfaces. This patch fixes that
by assuming there will be on entry point in DB if there
is no namespace and runs server and client on that endpoint.

Change-Id: I05fdc1e34e2bd2335a0ca1657af6f828bf740a84