Separate config and traffic DBs
Separate the traffic DB and config DB to separate commits.
This will help in storing traffic Data in different database
like mysql or other RDBMS which exists outside. And it also
separates the sqlite db file for both in the case sqlite is used.
It also adds indexing in trafficrecord table on fields created
and success, src and dst.

Jira: CX-1841

Change-Id: Ib3e5d715f3c0fa52b68366280c745b0d5f8ec7a2