Scale: Changes to traffic_elements, TrafficRecord source, retries
This commit adds class EndpointList to traffic_elements so that
we can provide a list of source or destination IPs (for later
support for creating rules that have no source or destination).
It also adds additional checks and assert messages to check
TrafficRule and provide feedback on invalid TrafficRule. It also
adds any traffic exception to the wavefront failure record. In
addition, it adds retries for the TCP, UDP, and HTTP protocols.
Finally, it changes wavefront TrafficRecord source to be the source IP
instead of hardcoded 'localhost'.

Change-Id: Idd7bc755c5849bfc800fd8e5643597052d23eb72