Scale: Record subset of failures to Wavefront and added logging
This commit uses a new Axon config parameter called
WAVEFRONT_REPORT_PERC which defines how often a failure will be
recorded. At scale, we can have a large number of failures which
can result in Wavefront API timeouts. We also don't need a lot of
failures in Wavefront to triage most connectivity issues. The default
is to report 100% of the failures. Also enhanced logging of traffic
failures in axon.log so in case we still have a lot of faliure
records and we can't retrieve them from Wavefront, we have a backup
option of scraping the the axon.log files.

Change-Id: I753367c79934d60b8a9524704d65c5352a5b3b84