axon: Adds Console app.
Change includes :
    - Adds Console app.
    - Exposes Resource Monitor App.

Console App shall be the basis for all the apps that require to run
another tool or process as subprocess. In future, Console app can be
considered for exposing as well if the need arises but it is not done
by default as it can pause security risk as it can be used for running
some malicious commands at endpoints.

Signed-off-by: Vipin Sharma <sharmavipin@vmware.com>
Change-Id: I80bf7a8228586e57f04549b59fee4825a1b88b10